requirements.txt
pygame==2.5.2
requests==2.31.0
pillow==10.2.0